refactor: extract worfklow steps as components
