+++ config-ui/src/pages/blueprints/create-blueprint.jsx
@@ -81,6 +81,8 @@ import usePipelineValidation from '@/hooks/usePipelineValidation'
 import useConnectionValidation from '@/hooks/useConnectionValidation'
 import useJIRA from '@/hooks/useJIRA'
 
+import WorkflowStepsBar from '@/components/blueprints/WorkflowStepsBar'
+import WorkflowActions from '@/components/blueprints/WorkflowActions'
 import FormValidationErrors from '@/components/messages/FormValidationErrors'
 import InputValidationError from '@/components/validation/InputValidationError'
 import ConnectionsSelector from '@/components/blueprints/ConnectionsSelector'
@@ -90,6 +92,11 @@ import ConnectionDialog from '@/components/blueprints/ConnectionDialog'
 import StandardStackedList from '@/components/blueprints/StandardStackedList'
 import CodeInspector from '@/components/pipelines/CodeInspector'
 import NoData from '@/components/NoData'
+
+import DataConnections from '@/components/blueprints/create-workflow/DataConnections'
+import DataScopes from '@/components/blueprints/create-workflow/DataScopes'
+import DataTransformations from '@/components/blueprints/create-workflow/DataTransformations'
+import DataSync from '@/components/blueprints/create-workflow/DataSync'
 
 import ConnectionTabs from '@/components/blueprints/ConnectionTabs'
 import ClearButton from '@/components/ClearButton'
@@ -548,7 +555,7 @@ const CreateBlueprint = (props) => {
     setConnectionDialogIsOpen(true)
   }
 
-  const renderProviderSettings = useCallback(
+  {/* const renderProviderSettings = useCallback(
     (activeConnection, providerId, activeProvider, dataEntityType) => {
       console.log('>>> RENDERING PROVIDER SETTINGS...')
       let settingsComponent = null
@@ -570,7 +577,7 @@ const CreateBlueprint = (props) => {
       return settingsComponent
     },
     [isSaving, isSavingConnection]
-  )
+  ) */}
 
   useEffect(() => {
     validateBlueprint()
@@ -673,886 +680,91 @@ const CreateBlueprint = (props) => {
         <Sidebar />
         <Content>
           <main className='main'>
-            <div className='workflow-bar'>
-              <ul className='workflow-steps'>
-                <li
-                  className={`workflow-step ${
-                    activeStep?.id === 1 ? 'active' : ''
-                  }`}
-                >
-                  <a href='#' className='step-id'>
-                    1
-                  </a>
-                  Add Data Connections
-                </li>
-                <li
-                  className={`workflow-step ${
-                    activeStep?.id === 2 ? 'active' : ''
-                  }`}
-                >
-                  <a href='#' className='step-id'>
-                    2
-                  </a>
-                  Set Data Scope
-                </li>
-                <li
-                  className={`workflow-step ${
-                    activeStep?.id === 3 ? 'active' : ''
-                  }`}
-                >
-                  <a href='#' className='step-id'>
-                    3
-                  </a>
-                  Add Transformation (Optional)
-                </li>
-                <li
-                  className={`workflow-step ${
-                    activeStep?.id === 4 ? 'active' : ''
-                  }`}
-                >
-                  <a href='#' className='step-id'>
-                    4
-                  </a>
-                  Set Sync Frequency
-                </li>
-              </ul>
-            </div>
+            <WorkflowStepsBar activeStep={activeStep} />
 
             <div
               className={`workflow-content workflow-step-id-${activeStep?.id}`}
             >
               {activeStep?.id === 1 && (
-                <div className='workflow-step workflow-step-data-connections'>
-                  <Card
-                    className='workflow-card'
-                    elevation={Elevation.TWO}
-                    style={{ width: '100%' }}
-                  >
-                    <h3>
-                      Blueprint Name <span className='required-star'>*</span>
-                    </h3>
-                    <Divider className='section-divider' />
-                    <p>
-                      Give your Blueprint a unique name to help you identify it
-                      in the future.
-                    </p>
-                    <InputGroup
-                      id='blueprint-name'
-                      placeholder='Enter Blueprint Name'
-                      value={name}
-                      onChange={(e) => setBlueprintName(e.target.value)}
-                      className={`blueprint-name-input ${
-                        fieldHasError('Blueprint Name') ? 'invalid-field' : ''
-                      }`}
-                      inline={true}
-                      style={{ marginBottom: '10px' }}
-                      rightElement={
-                        <InputValidationError
-                          error={getFieldError('Blueprint Name')}
-                        />
-                      }
-                    />
-                  </Card>
-
-                  <Card
-                    className='workflow-card'
-                    elevation={Elevation.TWO}
-                    style={{ width: '100%' }}
-                  >
-                    <div
-                      style={{
-                        display: 'flex',
-                        justifyContent: 'space-between',
-                      }}
-                    >
-                      <h3 style={{ margin: 0 }}>
-                        Add Data Connections{' '}
-                        <span className='required-star'>*</span>
-                      </h3>
-                      <div>
-                        <Button
-                          text='Add Connection'
-                          icon='plus'
-                          intent={Intent.PRIMARY}
-                          small
-                          onClick={addConnection}
-                        />
-                      </div>
-                    </div>
-                    <Divider className='section-divider' />
-
-                    <h4>Select Connections</h4>
-                    <p>Select from existing or create new connections</p>
-
-                    <ConnectionsSelector
-                      items={connectionsList}
-                      selectedItems={blueprintConnections}
-                      onItemSelect={setBlueprintConnections}
-                      onClear={setBlueprintConnections}
-                      onRemove={setBlueprintConnections}
-                      disabled={isSaving}
-                    />
-                    {blueprintConnections.length > 0 && (
-                      <Card
-                        className='selected-connections-list'
-                        elevation={Elevation.ZERO}
-                        style={{ padding: 0, marginTop: '10px' }}
-                      >
-                        {blueprintConnections.map((bC, bcIdx) => (
-                          <div
-                            className='connection-entry'
-                            key={`connection-row-key-${bcIdx}`}
-                            style={{
-                              display: 'flex',
-                              width: '100%',
-                              height: '32px',
-                              lineHeight: '100%',
-                              justifyContent: 'space-between',
-                              // margin: '8px 0',
-                              padding: '8px 12px',
-                              borderBottom: '1px solid #f0f0f0',
-                            }}
-                          >
-                            <div>
-                              <div
-                                className='connection-name'
-                                style={{ fontWeight: 600 }}
-                              >
-                                {bC.title}
-                              </div>
-                            </div>
-                            <div
-                              style={{
-                                display: 'flex',
-                                alignContent: 'center',
-                              }}
-                            >
-                              <div
-                                className='connection-status'
-                                style={{ textTransform: 'capitalize' }}
-                              >
-                                {bC.status}
-                              </div>
-                              <div
-                                className='connection-actions'
-                                style={{ paddingLeft: '20px' }}
-                              >
-                                <Button
-                                  className='connection-action-settings'
-                                  icon={
-                                    <Icon
-                                      icon='cog'
-                                      size={12}
-                                      color={Colors.BLUE4}
-                                      onClick={() => manageConnection(bC)}
-                                    />
-                                  }
-                                  color={Colors.BLUE3}
-                                  small
-                                  minimal
-                                  style={{
-                                    minWidth: '18px',
-                                    minHeight: '18px',
-                                  }}
-                                />
-                              </div>
-                            </div>
-                          </div>
-                        ))}
-                      </Card>
-                    )}
-                  </Card>
-                </div>
+                <DataConnections
+                  activeStep={activeStep}
+                  blueprintConnections={blueprintConnections}
+                  connectionsList={connectionsList}
+                  setBlueprintName={setBlueprintName}
+                  setBlueprintConnections={setBlueprintConnections}
+                  fieldHasError={fieldHasError}
+                  getFieldError={getFieldError}
+                  addConnection={addConnection}
+                  manageConnection={manageConnection}
+                />
               )}
 
               {activeStep?.id === 2 && (
-                <div className='workflow-step workflow-step-data-scope'>
-                  {blueprintConnections.length > 0 && (
-                    <div style={{ display: 'flex' }}>
-                      <div
-                        className='connection-tab-selector'
-                        style={{ minWidth: '200px' }}
-                      >
-                        <Card
-                          className='workflow-card connection-tabs-card'
-                          elevation={Elevation.TWO}
-                          style={{ padding: '10px' }}
-                        >
-                          <ConnectionTabs
-                            connections={blueprintConnections}
-                            onChange={handleConnectionTabChange}
-                            selectedTabId={activeConnectionTab}
-                          />
-                          {/*                         <Tabs
-                          className='connection-tabs'
-                          animate={true}
-                          vertical={true}
-                          id='connection-tabs'
-                          onChange={(tabId) => handleConnectionTabChange(tabId)}
-                          selectedTabId={activeConnectionTab}
-                        >
-                          {blueprintConnections.map((bC, bcIdx) => (
-                            <Tab
-                              key={`tab-key-${bcIdx}`}
-                              id={`connection-${bC.id}`}
-                              title={bC.title}
-                              connection={bC}
-                              // disabled={bC.disabled || bC.status !== 'online'}
-                            />
-                          ))}
-                        </Tabs> */}
-                        </Card>
-                      </div>
-                      <div
-                        className='connection-scope'
-                        style={{ marginLeft: '10px', width: '100%' }}
-                      >
-                        <Card
-                          className='workflow-card worfklow-panel-card'
-                          elevation={Elevation.TWO}
-                        >
-                          {configuredConnection && (
-                            <>
-                              <h3>
-                                <span
-                                  style={{ float: 'left', marginRight: '8px' }}
-                                >
-                                  {ProviderIcons[
-                                    configuredConnection.provider
-                                  ] ? (
-                                    ProviderIcons[
-                                      configuredConnection.provider
-                                    ](24, 24)
-                                  ) : (
-                                    <></>
-                                  )}
-                                </span>{' '}
-                                {configuredConnection.title}
-                              </h3>
-                              <Divider className='section-divider' />
-
-                              {[Providers.GITLAB, Providers.GITHUB].includes(
-                                configuredConnection.provider
-                              ) && (
-                                <>
-                                  <h4>Projects *</h4>
-                                  <p>
-                                    Enter the projects you would like to sync.
-                                  </p>
-                                  <TagInput
-                                    id='project-id'
-                                    disabled={isRunning}
-                                    placeholder={
-                                      configuredConnection.provider ===
-                                      Providers.GITHUB
-                                        ? 'username/repo, username/another-repo'
-                                        : '1000000, 200000'
-                                    }
-                                    values={
-                                      projects[configuredConnection.id] || []
-                                    }
-                                    fill={true}
-                                    onChange={(values) =>
-                                      setProjects((p) => ({
-                                        ...p,
-                                        [configuredConnection.id]: [
-                                          ...new Set(values),
-                                        ],
-                                      }))
-                                    }
-                                    addOnPaste={true}
-                                    addOnBlur={true}
-                                    rightElement={
-                                      <Button
-                                        disabled={isRunning}
-                                        icon='eraser'
-                                        minimal
-                                        onClick={() =>
-                                          setProjects((p) => ({
-                                            ...p,
-                                            [configuredConnection.id]: [],
-                                          }))
-                                        }
-                                      />
-                                    }
-                                    onKeyDown={(e) =>
-                                      e.key === 'Enter' && e.preventDefault()
-                                    }
-                                    tagProps={{
-                                      intent: Intent.PRIMARY,
-                                      minimal: true,
-                                    }}
-                                    className='input-project-id tagInput'
-                                  />
-                                </>
-                              )}
-
-                              {[Providers.JIRA].includes(
-                                configuredConnection.provider
-                              ) && (
-                                <>
-                                  <h4>Boards *</h4>
-                                  <p>
-                                    Select the boards you would like to sync.
-                                  </p>
-                                  <BoardsSelector
-                                    items={boardsList}
-                                    selectedItems={
-                                      boards[configuredConnection.id] || []
-                                    }
-                                    onItemSelect={setBoards}
-                                    onClear={setBoards}
-                                    onRemove={setBoards}
-                                    disabled={isSaving}
-                                    configuredConnection={configuredConnection}
-                                  />
-                                </>
-                              )}
-
-                              <h4>Data Entities</h4>
-                              <p>
-                                Select the data entities you wish to collect for
-                                the projects.{' '}
-                                <a href='#'>Learn about data entities</a>
-                              </p>
-                              <DataEntitiesSelector
-                                items={dataEntitiesList}
-                                selectedItems={
-                                  dataEntities[configuredConnection.id] || []
-                                }
-                                // restrictedItems={getRestrictedDataEntities()}
-                                onItemSelect={setDataEntities}
-                                onClear={setDataEntities}
-                                onRemove={setDataEntities}
-                                disabled={isSaving}
-                                configuredConnection={configuredConnection}
-                              />
-                            </>
-                          )}
-                        </Card>
-                      </div>
-                    </div>
-                  )}
-                  {blueprintConnections.length === 0 && (
-                    <NoData
-                      title='No Data Connections'
-                      message='Please select at least one connection source.'
-                      onClick={prevStep}
-                    />
-                  )}
-                </div>
+                <DataScopes
+                  activeStep={activeStep}
+                  activeConnectionTab={activeConnectionTab}
+                  blueprintConnections={blueprintConnections}
+                  dataEntitiesList={dataEntitiesList}
+                  dataEntities={dataEntities}
+                  projects={projects}
+                  configuredConnection={configuredConnection}
+                  handleConnectionTabChange={handleConnectionTabChange}
+                  setDataEntities={setDataEntities}
+                  setProjects={setProjects}
+                  prevStep={prevStep}
+                  isSaving={isSaving}
+                  isRunning={isRunning}
+                />
               )}
 
               {activeStep?.id === 3 && (
-                <div className='workflow-step workflow-step-add-transformation'>
-                  <p
-                    className='alert-neutral'
-                    style={{
-                      fontFamily: 'Montserrat, sans-serif',
-                      fontSize: '12px',
-                      margin: '0 0 10px 0',
-                      padding: '8px',
-                      borderRadius: '6px',
-                      color: '#3C5088',
-                      backgroundColor: '#F0F4FE',
-                      border: '1px solid #BDCEFB',
-                    }}
-                  >
-                    Set transformation rules for your selected data to view more
-                    complex metrics in the dashboards.
-                    <br />
-                    <a
-                      href='#'
-                      className='more-link'
-                      rel='noreferrer'
-                      style={{
-                        color: '#7497F7',
-                        marginTop: '5px',
-                        display: 'inline-block',
-                      }}
-                    >
-                      Find out more
-                    </a>
-                  </p>
-                  {blueprintConnections.length > 0 && (
-                    <div style={{ display: 'flex' }}>
-                      <div
-                        className='connection-tab-selector'
-                        style={{ minWidth: '200px' }}
-                      >
-                        <Card
-                          className='workflow-card connection-tabs-card'
-                          elevation={Elevation.TWO}
-                          style={{ padding: '10px' }}
-                        >
-                          <ConnectionTabs
-                            connections={blueprintConnections}
-                            onChange={handleConnectionTabChange}
-                            selectedTabId={activeConnectionTab}
-                          />
-                        </Card>
-                      </div>
-                      <div
-                        className='connection-transformation'
-                        style={{ marginLeft: '10px', width: '100%' }}
-                      >
-                        <Card
-                          className='workflow-card workflow-panel-card'
-                          elevation={Elevation.TWO}
-                        >
-                          {configuredConnection && (
-                            <>
-                              <h3>
-                                <span
-                                  style={{ float: 'left', marginRight: '8px' }}
-                                >
-                                  {ProviderIcons[
-                                    configuredConnection.provider
-                                  ] ? (
-                                    ProviderIcons[
-                                      configuredConnection.provider
-                                    ](24, 24)
-                                  ) : (
-                                    <></>
-                                  )}
-                                </span>{' '}
-                                {configuredConnection.title}
-                              </h3>
-                              <Divider className='section-divider' />
-
-                              {[Providers.GITLAB, Providers.GITHUB].includes(
-                                configuredConnection.provider
-                              ) && (
-                                <>
-                                  <StandardStackedList
-                                    items={projects}
-                                    className='selected-items-list selected-projects-list'
-                                    connection={configuredConnection}
-                                    activeItem={configuredProject}
-                                    onAdd={addProjectTransformation}
-                                    onChange={addProjectTransformation}
-                                  />
-                                  {projects[configuredConnection.id].length ===
-                                    0 && (
-                                    <NoData
-                                      title='No Projects Selected'
-                                      icon='git-branch'
-                                      message='Please select specify at least one project.'
-                                      onClick={prevStep}
-                                    />
-                                  )}
-                                </>
-                              )}
-
-                              {[Providers.JIRA].includes(
-                                configuredConnection.provider
-                              ) && (
-                                <>
-                                  <StandardStackedList
-                                    items={boards}
-                                    className='selected-items-list selected-boards-list'
-                                    connection={configuredConnection}
-                                    activeItem={configuredBoard}
-                                    onAdd={addBoardTransformation}
-                                    onChange={addBoardTransformation}
-                                  />
-                                  {boards[configuredConnection.id].length ===
-                                    0 && (
-                                    <NoData
-                                      title='No Boards Selected'
-                                      icon='th'
-                                      message='Please select specify at least one board.'
-                                      onClick={prevStep}
-                                    />
-                                  )}
-                                </>
-                              )}
-
-                              {configuredProject && (
-                                <>
-                                  <h4>Project</h4>
-                                  <p>
-                                    {configuredProject ||
-                                      '< select a project >'}
-                                  </p>
-                                  <h4>Data Transformation Rules</h4>
-                                  {!dataEntities[configuredConnection.id] ||
-                                    (dataEntities[configuredConnection.id]
-                                      ?.length === 0 && (
-                                      <p>(No Data Entities Selected)</p>
-                                    ))}
-                                  {dataEntities[configuredConnection.id]?.find(
-                                    (e) => e.value === DataEntityTypes.TICKET
-                                  ) && (
-                                    <>
-                                      {renderProviderSettings(
-                                        configuredConnection,
-                                        configuredConnection.provider,
-                                        integrationsData.find(
-                                          (p) =>
-                                            p.id ===
-                                            configuredConnection.provider
-                                        ),
-                                        DataEntityTypes.TICKET
-                                      )}
-                                    </>
-                                  )}
-                                </>
-                              )}
-                            </>
-                          )}
-                        </Card>
-                      </div>
-                    </div>
-                  )}
-                  {blueprintConnections.length === 0 && (
-                    <>
-                      <div className='bp3-non-ideal-state'>
-                        <div className='bp3-non-ideal-state-visual'>
-                          <Icon icon='offline' size={32} />
-                        </div>
-                        <div className='bp3-non-ideal-state-text'>
-                          <h4 className='bp3-heading' style={{ margin: 0 }}>
-                            No Data Connections
-                          </h4>
-                          <div>
-                            Please select at least one connection source.
-                          </div>
-                        </div>
-                        <button
-                          className='bp3-button bp4-intent-primary'
-                          onClick={prevStep}
-                        >
-                          Go Back
-                        </button>
-                      </div>
-                    </>
-                  )}
-                </div>
+                <DataTransformations
+                  activeProvider={activeProvider}
+                  activeStep={activeStep}
+                  activeConnectionTab={activeConnectionTab}
+                  blueprintConnections={blueprintConnections}
+                  dataEntities={dataEntities}
+                  projects={projects}
+                  boards={boards}
+                  configuredConnection={configuredConnection}
+                  configuredProject={configuredProject}
+                  handleConnectionTabChange={handleConnectionTabChange}
+                  prevStep={prevStep}
+                  addBoardTransformation={addBoardTransformation}
+                  addProjectTransformation={addProjectTransformation}
+                  // renderProviderSettings={renderProviderSettings}
+                  isSaving={isSaving}
+                  isSavingConnection={isSavingConnection}
+                  isRunning={isRunning}
+                />
               )}
 
               {activeStep?.id === 4 && (
-                <div className='workflow-step workflow-step-set-sync-frequency'>
-                  <Card className='workflow-card' elevation={Elevation.TWO}>
-                    <h3 style={{ marginBottom: '8px' }}>Set Sync Frequency</h3>
-                    {getCronPresetByConfig(cronConfig) ? (
-                      <small
-                        style={{
-                          fontSize: '10px',
-                          color: Colors.GRAY2,
-                          display: 'block',
-                        }}
-                      >
-                        <strong>Automated</strong> &mdash;{' '}
-                        {getCronPresetByConfig(cronConfig).description}
-                      </small>
-                    ) : (
-                      <small
-                        style={{
-                          fontSize: '10px',
-                          color: Colors.GRAY2,
-                          textTransform: 'uppercase',
-                        }}
-                      >
-                        {cronConfig}
-                      </small>
-                    )}
-                    <Divider className='section-divider' />
-
-                    <h4>Frequency</h4>
-                    <p>
-                      Blueprints will run recurringly based on the sync
-                      frequency.
-                    </p>
-
-                    <RadioGroup
-                      inline={false}
-                      label={false}
-                      name='blueprint-frequency'
-                      onChange={(e) => setCronConfig(e.target.value)}
-                      selectedValue={cronConfig}
-                      required
-                    >
-                      <Radio
-                        label='Manual'
-                        value='manual'
-                        style={{
-                          fontWeight:
-                            cronConfig === 'manual' ? 'bold' : 'normal',
-                        }}
-                      />
-                      {/* Dynamic Presets from Connection Manager */}
-                      {[
-                        getCronPreset('hourly'),
-                        getCronPreset('daily'),
-                        getCronPreset('weekly'),
-                        getCronPreset('monthly'),
-                      ].map((preset, prIdx) => (
-                        <Radio
-                          key={`cron-preset-tooltip-key${prIdx}`}
-                          label={
-                            <>
-                              <Tooltip
-                                position={Position.RIGHT}
-                                intent={Intent.PRIMARY}
-                                content={preset.description}
-                              >
-                                {preset.label}
-                              </Tooltip>
-                            </>
-                          }
-                          value={preset.cronConfig}
-                          style={{
-                            fontWeight:
-                              cronConfig === preset.cronConfig
-                                ? 'bold'
-                                : 'normal',
-                            outline: 'none !important',
-                          }}
-                        />
-                      ))}
-                      <Radio
-                        label='Custom'
-                        value='custom'
-                        style={{
-                          fontWeight:
-                            cronConfig === 'custom' ? 'bold' : 'normal',
-                        }}
-                      />
-                    </RadioGroup>
-                    <div
-                      style={{
-                        display: cronConfig === 'custom' ? 'flex' : 'none',
-                      }}
-                    >
-                      <FormGroup
-                        disabled={cronConfig !== 'custom'}
-                        label=''
-                        inline={true}
-                        labelFor='cron-custom'
-                        className='formGroup-inline'
-                        contentClassName='formGroupContent'
-                        style={{ marginBottom: '5px' }}
-                        fill={false}
-                      >
-                        <InputGroup
-                          id='cron-custom'
-                          inline={true}
-                          fill={false}
-                          readOnly={cronConfig !== 'custom'}
-                          leftElement={
-                            cronConfig !== 'custom' ? (
-                              <Icon
-                                icon='lock'
-                                size={11}
-                                style={{
-                                  alignSelf: 'center',
-                                  margin: '4px 10px -2px 6px',
-                                }}
-                              />
-                            ) : null
-                          }
-                          rightElement={
-                            <>
-                              <InputValidationError
-                                error={getFieldError('Blueprint Cron')}
-                              />
-                            </>
-                          }
-                          placeholder='Enter Crontab Syntax'
-                          value={
-                            cronConfig !== 'custom'
-                              ? cronConfig
-                              : customCronConfig
-                          }
-                          onChange={(e) => setCustomCronConfig(e.target.value)}
-                          className={`cron-custom-input ${
-                            fieldHasError('Blueprint Cron')
-                              ? 'invalid-field'
-                              : ''
-                          }`}
-                          inline={true}
-                          fill={false}
-                          style={{ transition: 'none' }}
-                        />
-                      </FormGroup>
-                      <div
-                        style={{
-                          display: 'inline',
-                          marginTop: 'auto',
-                          paddingBottom: '15px',
-                        }}
-                      >
-                        <Popover
-                          className='trigger-crontab-help'
-                          popoverClassName='popover-help-crontab'
-                          position={Position.RIGHT}
-                          autoFocus={false}
-                          enforceFocus={false}
-                          usePortal={false}
-                        >
-                          <a rel='noreferrer'>
-                            <Icon
-                              icon='help'
-                              size={14}
-                              style={{ marginLeft: '10px', transition: 'none' }}
-                            />
-                          </a>
-                          <>
-                            <div
-                              style={{
-                                textShadow: 'none',
-                                fontSize: '12px',
-                                padding: '12px',
-                                maxWidth: '300px',
-                              }}
-                            >
-                              <div
-                                style={{
-                                  marginBottom: '10px',
-                                  fontWeight: 700,
-                                  fontSize: '14px',
-                                  fontFamily: '"Montserrat", sans-serif',
-                                }}
-                              >
-                                <Icon
-                                  icon='help'
-                                  size={16}
-                                  style={{ marginRight: '5px' }}
-                                />{' '}
-                                Cron Expression Format
-                              </div>
-                              <p>
-                                Need Help? &mdash; For additional information on{' '}
-                                <strong>Crontab</strong>, please reference the{' '}
-                                <a
-                                  href='https://man7.org/linux/man-pages/man5/crontab.5.html'
-                                  rel='noreferrer'
-                                  target='_blank'
-                                  style={{ textDecoration: 'underline' }}
-                                >
-                                  Crontab Linux manual
-                                </a>
-                                .
-                              </p>
-                              <img
-                                src={CronHelp}
-                                style={{
-                                  border: 0,
-                                  margin: 0,
-                                  maxWidth: '100%',
-                                }}
-                              />
-                            </div>
-                          </>
-                        </Popover>
-                      </div>
-                    </div>
-
-                    {cronConfig !== 'manual' && (
-                      <div>
-                        <Divider
-                          className='section-divider'
-                          style={{ marginTop: ' 20px' }}
-                        />
-                        <div>
-                          <Button
-                            text='View Schedule'
-                            icon='time'
-                            intent={Intent.NONE}
-                            small
-                            style={{ float: 'right', fontSize: '11px' }}
-                          />
-
-                          <h4 style={{ marginRight: 0, marginBottom: 0 }}>
-                            Next Run Date
-                          </h4>
-                        </div>
-                        <div style={{ fontSize: '14px', fontWeight: 800 }}>
-                          {dayjs(
-                            createCron(
-                              cronConfig === 'custom'
-                                ? customCronConfig
-                                : cronConfig
-                            )
-                              .next()
-                              .toString()
-                          ).format('L LTS')}{' '}
-                          &middot;{' '}
-                          <span style={{ color: Colors.GRAY3 }}>
-                            (
-                            {dayjs(
-                              createCron(
-                                cronConfig === 'custom'
-                                  ? customCronConfig
-                                  : cronConfig
-                              )
-                                .next()
-                                .toString()
-                            ).fromNow()}
-                            )
-                          </span>
-                        </div>
-                      </div>
-                    )}
-                  </Card>
-                </div>
+                <DataSync
+                  cronConfig={cronConfig}
+                  customCronConfig={customCronConfig}
+                  createCron={createCron}
+                  setCronConfig={setCronConfig}
+                  getCronPreset={getCronPreset}
+                  fieldHasError={fieldHasError}
+                  getFieldError={getFieldError}
+                  setCustomCronConfig={setCustomCronConfig}
+                  getCronPresetByConfig={getCronPresetByConfig}
+                />
               )}
             </div>
 
-            <div className='workflow-actions'>
-              <Button
-                intent={Intent.PRIMARY}
-                text='Previous Step'
-                onClick={prevStep}
-                disabled={activeStep?.id === 1}
-              />
-
-              {activeStep?.id === 4 ? (
-                <div style={{ marginLeft: 'auto' }}>
-                  <Button
-                    intent={Intent.PRIMARY}
-                    text='Save Blueprint'
-                    disabled
-                    // onClick={saveBlueprint}
-                  />
-                  <Button
-                    intent={Intent.DANGER}
-                    text='Save and Run Now'
-                    style={{ marginLeft: '5px' }}
-                    disabled
-                    // onClick={saveBlueprintAndRunPipeline}
-                  />
-                </div>
-              ) : (
-                <div style={{ display: 'flex', marginLeft: 'auto' }}>
-                  <Button
-                    intent={Intent.PRIMARY}
-                    icon='code'
-                    text='Inspect'
-                    onClick={() => setShowBlueprintInspector(true)}
-                    style={{ marginRight: '8px' }}
-                    minimal
-                    small
-                  />
-                  <Button
-                    intent={Intent.PRIMARY}
-                    text='Next Step'
-                    onClick={nextStep}
-                    rightIcon={
-                        validationErrors.length > 0 ? (
-                          <Popover interactionKind={PopoverInteractionKind.HOVER_TARGET_ONLY}>
-                            <Icon icon='warning-sign' size={12} color={Colors.ORANGE5} style={{ outline: 'none', margin: '0 3px 2px 3px' }} />
-                            <div style={{ padding: '5px' }}><FormValidationErrors errors={validationErrors} /></div>
-                          </Popover>
-                        ) : null}
-                  />
-                </div>
-              )}
-            </div>
+            <WorkflowActions 
+              activeStep={activeStep}
+              setShowBlueprintInspector={setShowBlueprintInspector}
+              validationErrors={validationErrors}
+              nextStep={nextStep}
+              prevStep={prevStep}
+            />
           </main>
         </Content>
       </div>
+      
       <ConnectionDialog
         integrations={integrationsData}
         activeProvider={activeProvider}
